+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
         android:theme="@style/ThemeApplication"
         tools:targetApi="31">
         <receiver
-            android:name=".service.AlarmReceiver"
+            android:name=".receiver.AlarmReceiver"
             android:exported="false">
             <intent-filter>
                 <action android:name="alarm.ACTION_DISMISS" />
